fix: change currentCorrectOptionIndex to Short in QuizSessionJpaEntity

+++ quiz-service/src/main/java/uz/quizplatform/quizservice/infrastructure/persistence/entity/QuizSessionJpaEntity.java
@@ -63,7 +63,7 @@ public class QuizSessionJpaEntity {
     private String currentPollId;
 
     @Column(name = "current_correct_option_idx")
-    private Integer currentCorrectOptionIndex;
+    private Short currentCorrectOptionIndex;
 
     @Column(name = "current_question_sent_at")
     private Instant currentQuestionSentAt;
